Diamond square implemented breadth-first, colors set to normals

+++ project/Terrain.h
@@ -27,7 +27,9 @@ private:
 	// 1D Midpoint displacement algorithm, takes line segment
 	// (x1, y1, x2, y2) and initial random range & roughness const
 	void midpointDisplacement(int, int, int, int, double, double, int);
-	void diamondSquare(int, int, int, int, double, double);
+	// void diamondSquare(int, int, int, int, double, double);
+
+	void diamondSquare(int, int, int, int, double, int); // iterative taken from SO
 
 public:
 	Terrain();
